Split core
