+++ ffmpegwithpy-rust/Cargo.toml
@@ -13,6 +13,7 @@ crate-type = ["cdylib"]
 
 [dependencies]
 pyo3 = { version = "0.17.3", features = ["extension-module", "abi3-py37"] }
+ffmpegwithpy-core = { path = "../ffmpegwithpy-core"}
 
 [package.metadata.maturin]
 python-source = "python"
